vello_hybrid: Extract probing code into its module
